+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Remove unused callback from tests
 - Add loop_safe annotation
+- Remove entity_data_event_callback
 
 # Version 2024.4.5 (2024-04-09)